+++ R/specreator.R
@@ -89,7 +89,7 @@
 #' 
 #' specreator(selec.table, flim = c(0, 11), res = 300, mar = 0.05, wl = 300)
 #'  
-#'  #' #check this folder!!
+#'  #check this folder
 #' getwd()
 #' }
 #'