+++ NEWS.md
@@ -15,6 +15,9 @@
 * frange.detec: detects the frequency range of signals in wave objects
 * frange: applies 'frange.detec' function iteratively on signals listed in a selection table
 * move.imgs: copy/cut image files to folders
+* ovlp.sels: find overlapping/duplicated selections
+* cut.sels: generates individual sound files (.wav) for each selection in a selection table
+
 ===============================================================================================
 
 # warbleR 1.1.8